refactor(rust): manage build via `BuildFactory` for `Bundler` (#6883)
Follow-up of https://github.com/rolldown/rolldown/pull/6877.

This PR makes the native `Bundler`​ integrate with the `Build`​ abstraction too.
